ProjectExplorer: Fix always asking to remove files from VCS
Regardless the "Delete file permanently" checkbox state.

If the checkbox is not checked, we should just remove the
file from the project and leave the VCS and disk state
untouched.

Most likely introduced by commit
3b415f5ebd4936b9fdf2d2b211ee368d7f943fa7

Fixes: QTCREATORBUG-32856

+++ src/plugins/projectexplorer/projectexplorer.cpp
@@ -3912,7 +3912,8 @@ void ProjectExplorerPluginPrivate::removeFile()
         changeGuards.emplace_back(std::make_unique<FileChangeBlocker>(file.second));
     }
 
-    Core::FileUtils::removeFiles(pathList, deleteFile);
+    if (deleteFile)
+        Core::FileUtils::removeFiles(pathList, deleteFile);
 }
 
 void ProjectExplorerPluginPrivate::duplicateFile()
